Move core::math_utils helpers into bezier::utils and tessellator::math_utils.
