+++ tessellation/src/path_fill.rs
@@ -125,7 +125,8 @@ use Side;
 use math::*;
 use geometry_builder::{GeometryBuilder, Count, VertexId};
 use core::FlattenedEvent;
-use math_utils::{directed_angle, directed_angle2};
+use bezier::utils::{directed_angle, directed_angle2};
+use math_utils::{line_horizontal_intersection_fixed, segment_intersection};
 
 #[cfg(test)]
 use geometry_builder::{VertexBuffers, simple_builder};
@@ -1116,99 +1117,6 @@ fn compare_positions(a: TessPoint, b: TessPoint) -> Ordering {
         return Ordering::Less;
     }
     return Ordering::Equal;
-}
-
-fn line_horizontal_intersection_fixed(
-    a: TessPoint,
-    b: TessPoint,
-    y: FixedPoint32,
-) -> Option<FixedPoint32> {
-    let vx = b.x - a.x;
-    let vy = b.y - a.y;
-
-    if vy.is_zero() {
-        // the line is horizontal
-        return None;
-    }
-
-    let tmp: FixedPoint64 = (y - a.y).to_fp64();
-    return Some(a.x + tmp.mul_div(vx.to_fp64(), vy.to_fp64()).to_fp32());
-}
-
-#[derive(PartialEq, Debug)]
-enum SegmentInteresection {
-    One(TessPoint),
-    Two(TessPoint, TessPoint),
-    None,
-}
-
-#[inline]
-fn x_aabb_test(a1: FixedPoint32, b1: FixedPoint32, a2: FixedPoint32, b2: FixedPoint32) -> bool {
-    let (min1, max1) = a1.min_max(b1);
-    let (min2, max2) = a2.min_max(b2);
-    min1 < max2 && max1 > min2
-}
-
-// TODO[optim]: This function shows up pretty high in the profiles.
-fn segment_intersection(
-    a1: TessPoint,
-    b1: TessPoint, // The new edge.
-    a2: TessPoint,
-    b2: TessPoint, // An already inserted edge.
-) -> Option<TessPoint> {
-
-    // This early-out test gives a noticeable performance improvement.
-    if !x_aabb_test(a1.x, b1.x, a2.x, b2.x) {
-        return None;
-    }
-
-    fn tess_point(x: f64, y: f64) -> TessPoint {
-        TessPoint::new(FixedPoint32::from_f64(x), FixedPoint32::from_f64(y))
-    }
-
-    // TODO: moving this down after the v1_cross_v2 == 0.0 branch fixes test
-    // test_colinear_touching_squares2_failing
-    if a1 == b2 || a1 == a2 || b1 == a2 || b1 == b2 {
-        return None;
-    }
-
-    let a1 = F64Point::new(a1.x.to_f64(), a1.y.to_f64());
-    let b1 = F64Point::new(b1.x.to_f64(), b1.y.to_f64());
-    let a2 = F64Point::new(a2.x.to_f64(), a2.y.to_f64());
-    let b2 = F64Point::new(b2.x.to_f64(), b2.y.to_f64());
-
-    let v1 = b1 - a1;
-    let v2 = b2 - a2;
-
-    debug_assert!(v2.x != 0.0 || v2.y != 0.0, "zero-length edge");
-
-    let v1_cross_v2 = v1.cross(v2);
-    let a2_a1_cross_v1 = (a2 - a1).cross(v1);
-
-    if v1_cross_v2 == 0.0 {
-        return None;
-    }
-
-    let sign_v1_cross_v2 = if v1_cross_v2 > 0.0 { 1.0 } else { -1.0 };
-    let abs_v1_cross_v2 = v1_cross_v2 * sign_v1_cross_v2;
-
-    // t and u should be divided by v1_cross_v2, but we postpone that to not lose precision.
-    // We have to respect the sign of v1_cross_v2 (and therefore t and u) so we apply it now and
-    // will use the absolute value of v1_cross_v2 afterwards.
-    let t = (a2 - a1).cross(v2) * sign_v1_cross_v2;
-    let u = a2_a1_cross_v1 * sign_v1_cross_v2;
-
-    if t > 0.0 && t < abs_v1_cross_v2 && u > 0.0 && u <= abs_v1_cross_v2 {
-
-        let res = a1 + (v1 * t) / abs_v1_cross_v2;
-        debug_assert!(res.y <= b1.y && res.y <= b2.y);
-
-        if res != a1 && res != b1 && res != a2 && res != b2 {
-            return Some(tess_point(res.x, res.y));
-        }
-    }
-
-    return None;
 }
 
 // Returns true if the position is on the right side of the edge.
